+++ packages/ui-kit-js/src/types.d.ts
@@ -1,6 +1,5 @@
-// global text variants
-export type TextVariant = 'x-small' | 'small' | 'medium' | 'large' | 'x-large' | 'xx-large' | 'inherit';
+// global text sizes
+export type TextSize = 'x-small' | 'small' | 'medium' | 'large' | 'x-large' | 'inherit';
 
 // global text colors
 export type TextColor = 'porsche-black' | 'porsche-light' | 'porsche-red' | 'inherit';
-