+++ packages/ui-kit-js/CHANGELOG.md
@@ -7,7 +7,7 @@
 * Text component
   * Reduce amount of text variants in global interface
   * Remove `ul | ol | li | sup | sub | label` from `tag` property
-  * Add `thin` property
+  * Add `weight` property
   
 * Text Link component
   * Reduce amount of text variants (defined in global interface)